[refactor] 패키지 구조 변경 및 파일 이동
