+++ src/main/java/com/MooBoo/MooBoo_Spring/application/service/BookApiServiceImpl.java
@@ -1,6 +1,6 @@
 package com.MooBoo.MooBoo_Spring.application.service;
 
-import com.MooBoo.MooBoo_Spring.application.port.inbound.bookapi.BookApiService;
+import com.MooBoo.MooBoo_Spring.application.port.inbound.BookApiService;
 import com.MooBoo.MooBoo_Spring.domain.bookapi.BookApi;
 import com.MooBoo.MooBoo_Spring.adapter.inbound.api.bookapi.dto.SearchParam;
 import com.MooBoo.MooBoo_Spring.application.port.outbound.external.bookapi.BookApiClient;
